+++ Changelog.md
@@ -1,4 +1,5 @@
 # Version 1.0.0
+- Feature: Smart completion of instructions. Only those instructions which are compatible with the current stack are suggested.
 - Don't suggest section names inside of sections or types
 
 # Version 0.10.0